Pointers, arrays and strings

+++ 0x05-pointers_arrays_strings/7-puts_half.c
@@ -9,7 +9,7 @@
 
 void puts_half(char *str)
 {
-	int i, n;
+	unsigned long int i, n;
 
 	if (strlen(str) % 2 == 0)
 		n = (strlen(str) / 2);
